Fixed middleware route finding when using sub-apps (#65)
Previously, the middleware that gets installed references the current
requests app in order to locate the debugtoolbar url. However, if we
only setup the toolbar on the main app and then use
`app.router.add_subapp`, the subapps don't have access to these paths.

The fix is to get the path from the app instance that is availible in
the middleware's scope from when the setup first occured.